MCO2 v.4

+++ Fertilizer.java
@@ -1,11 +1,27 @@
+/**
+ * This class represents the fertilizer object that can be used by the player for the game.
+ * It contains attributes related to the cost of using the fertilizer and the experience 
+ * generated from using the plow
+ */
 public class Fertilizer extends Tool{
-    private int costFromUsage = 10;
+    private double costFromUsage = 10;
     private double experienceFromUsage = 4;
 
+    /**
+     * This constructor currently does not set any of the fertilizer's attributes to a default value
+     */
     public Fertilizer() {
 
     }
 
+    /**
+     * This method only activates when player has enough money to use fertilizer,
+     * represents the action of using the fertilizer on a tile,
+     * and can only be used when the tile has been plowed and has a crop
+     * 
+     * @param tile - the tile to be fertilized
+     * @return - returns true if the Fertilizer was successfully used
+     */
     public boolean useTool(Tile tile) {
         if(tile.getIsPlowed() && tile.getHasCrop()) {
             tile.getCrop().addTimesCropAddedFertilizer();
@@ -16,9 +32,19 @@ public class Fertilizer extends Tool{
             return false;
         }
     }
+
+    /**
+     * This returns the cost to be deducted from using the fertilizer
+     * 
+     * @return - the cost from using the fertilizer
+     */
     public double getCostFromUsage() {
         return this.costFromUsage;
     }
+
+    /**
+     * This returns the experienced to be added from using the fertilizer
+     */
     public double getExperienceFromUsage() {
         return this.experienceFromUsage;
     }
